backend/requirements.txt
# Python packages
google-adk
httpx
python-dotenv
litellm

# System dependencies (install separately)
# - Java 17:     brew install openjdk@17
# - Maven:       brew install maven
# - Node.js:     preinstalled in Daytona; brew install node for local deploy
# - Angular CLI: npm install -g @angular/cli@17
